requirements.txt
Flask>=3.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=2.1.4
openpyxl>=3.1.2
lxml>=4.9.3
python-dotenv>=1.0.0